✨ Feat #11: Suspense
Content:
- page.tsx에 있는 두 가지 data fetching을 components 폴더에 각각 movie-info.tsx, movie-info.tsx 파일로 분리했다.
- 각각의 파일들은 async await 컴포넌트를 가지고 있고 data fetching을 한다.
- page.tsx 파일은 이 컴포넌트들을 불러와 Suspense 컴포넌트로 감싸서 data fetching을 병렬처리 했다.
- Suspense 컴포넌트의 fallback prop은 data fetching이 일어나는 동안 보여줄 컴포넌트이다.
- loading.tsx: 전체 페이지를 로딩 상태로 대체
- Suspense Component: fetch해야 하는 component만 로딩 상태로 대체
