+++ components/movie-info.tsx
@@ -1,0 +1,14 @@
+import { API_URL } from '../app/(home)/page';
+
+async function getMovie(id: string) {
+  console.log(`Fetching movies: ${Date.now()}`);
+  //await new Promise((resolve) => setTimeout(resolve, 5000));
+
+  const response = await fetch(`${API_URL}/${id}`);
+  return response.json();
+}
+
+export default async function MovieInfo({ id }: { id: string }) {
+  const movies = await getMovie(id);
+  return <h6>{JSON.stringify(movies)}</h6>;
+}
